+++ NEWS.md
@@ -365,9 +365,13 @@ This section lists changes that do not have deprecation warnings.
   * `findn(x::AbstractVector)` now return a 1-tuple with the vector of indices, to be
     consistent with higher order arrays ([#25365]).
 
-  * the default behavior of `titlecase` is changed such that characters not starting
-    a word are converted to lowercase; a new keyword argument `strict` is added which
-    allows to get the old behavior when it's `false`.
+  * the default behavior of `titlecase` is changed in two ways ([#23393]):
+    + characters not starting a word are converted to lowercase;
+      a new keyword argument `strict` is added which
+      allows to get the old behavior when it's `false`.
+    + any non-letter character is considered as a word separator;
+      to get the old behavior (only "space" characters are considered as
+      word separators), use the keyword `wordsep=isspace`.
 
 
 Library improvements